+++ CHANGES.rst
@@ -6,6 +6,7 @@ Version 0.4.0 (UNRELEASED)
 
 - Adds support for Python 3.11.
 - Drops support for Python 2.7.
+- Improves ``get-metadata --output-field`` command by adding a new ``--filter`` option allowing to output only selected field values matching desired criteria.
 
 Version 0.3.0 (2022-03-16)
 --------------------------